HSEARCH-4089 Add a separate Maven profile to test with Elasticsearch 7.8+
