+++ integrationtest/backend/elasticsearch/pom.xml
@@ -200,9 +200,19 @@
             </properties>
         </profile>
 
-        <!-- Elasticsearch 7.7+ test environment (default) -->
+        <!-- Elasticsearch 7.7 test environment -->
         <profile>
             <id>elasticsearch-7.7</id>
+            <properties>
+                <failsafe.excludedGroups.elasticsearch.version>
+                    <!-- Nothing here -->
+                </failsafe.excludedGroups.elasticsearch.version>
+            </properties>
+        </profile>
+
+        <!-- Elasticsearch 7.8+ test environment (default) -->
+        <profile>
+            <id>elasticsearch-7.8</id>
             <activation>
                 <!-- This must be re-defined here: the activation defined in the parent pom is not enough -->
                 <!-- Activate by default, i.e. if test.elasticsearch.connection.version has not been defined explicitly -->
