+++ parents/integrationtest/pom.xml
@@ -478,17 +478,11 @@
             </build>
         </profile>
 
-        <!-- Elasticsearch 7.7+ test environment (default) -->
+        <!-- Elasticsearch 7.7 test environment -->
         <profile>
             <id>elasticsearch-7.7</id>
-            <activation>
-                <!-- Activate by default, i.e. if test.elasticsearch.connection.version has not been defined explicitly -->
-                <property>
-                    <name>!test.elasticsearch.connection.version</name>
-                </property>
-            </activation>
             <properties>
-                <test.elasticsearch.connection.version>${version.org.elasticsearch.main}</test.elasticsearch.connection.version>
+                <test.elasticsearch.connection.version>7.7.1</test.elasticsearch.connection.version>
                 <test.elasticsearch.testdialect>org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.Elasticsearch77TestDialect</test.elasticsearch.testdialect>
             </properties>
             <build>
@@ -501,6 +495,36 @@
                             <configuration>
                                 <pathConf>${project.build.directory}/elasticsearch-maven-plugin/7.0/configuration/</pathConf>
                                 <pathInitScript>${project.build.directory}/elasticsearch-maven-plugin/7.0/init/init.script</pathInitScript>
+                            </configuration>
+                        </plugin>
+                    </plugins>
+                </pluginManagement>
+            </build>
+        </profile>
+
+        <!-- Elasticsearch 7.8+ test environment (default) -->
+        <profile>
+            <id>elasticsearch-7.8</id>
+            <activation>
+                <!-- Activate by default, i.e. if test.elasticsearch.connection.version has not been defined explicitly -->
+                <property>
+                    <name>!test.elasticsearch.connection.version</name>
+                </property>
+            </activation>
+            <properties>
+                <test.elasticsearch.connection.version>${version.org.elasticsearch.main}</test.elasticsearch.connection.version>
+                <test.elasticsearch.testdialect>org.hibernate.search.util.impl.integrationtest.backend.elasticsearch.dialect.Elasticsearch78TestDialect</test.elasticsearch.testdialect>
+            </properties>
+            <build>
+                <pluginManagement>
+                    <plugins>
+                        <plugin>
+                            <groupId>com.github.alexcojocaru</groupId>
+                            <artifactId>elasticsearch-maven-plugin</artifactId>
+                            <version>${version.com.github.alexcojocaru.elasticsearch.plugin}</version>
+                            <configuration>
+                                <pathConf>${project.build.directory}/elasticsearch-maven-plugin/7.8/configuration/</pathConf>
+                                <pathInitScript>${project.build.directory}/elasticsearch-maven-plugin/7.8/init/init.script</pathInitScript>
                             </configuration>
                         </plugin>
                     </plugins>
