+++ util/internal/integrationtest/sharedresources/src/main/elasticsearch-maven-plugin/7.8/configuration/log4j2.properties
@@ -1,0 +1,36 @@
+status = error
+
+appender.console.type = Console
+appender.console.name = console
+appender.console.layout.type = PatternLayout
+appender.console.layout.pattern = [ES][%-5p][%d{ISO8601}][%-25c{1.}][%node_name]%marker %m%n
+
+rootLogger.level = info
+rootLogger.appenderRef.console.ref = console
+
+loggers = action, metadata, cluster, settings, deprecation, slow_search, slow_indexing
+
+# log action execution errors for easier debugging
+logger.action.name = org.elasticsearch.action
+logger.action.level = info
+
+# do not log metadata too much as we generate a log of noise
+logger.metadata.name = org.elasticsearch.cluster.metadata
+logger.metadata.level = warn
+
+logger.cluster.name = org.elasticsearch.cluster.routing.allocation
+logger.cluster.level = warn
+
+logger.settings.name = org.elasticsearch.common.settings
+logger.settings.level = warn
+
+logger.deprecation.name = org.elasticsearch.deprecation
+logger.deprecation.level = warn
+
+# Warn us about using inefficient search operations
+logger.slow_search.name = index.search.slowlog
+logger.slow_search.level = trace
+
+# Warn us about using inefficient indexing actions
+logger.slow_indexing.name = index.indexing.slowlog
+logger.slow_indexing.level = trace
